requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt

azure-servicebus==0.21.0
# pyinotify==0.9.6